Casting cases from bigint to int at DB level
